+++ com.unity.mobile.notifications/Runtime/Android/AndroidNotificationCenter.cs
@@ -170,6 +170,11 @@ namespace Unity.Notifications.Android
             return klass.CallStatic<string>(getNotificationChannelId, notification);
         }
 
+        public void RegisterNotificationChannelGroup(AndroidNotificationChannelGroup group)
+        {
+            self.Call("registerNotificationChannelGroup", group.Id, group.Name, group.Description);
+        }
+
         public void RegisterNotificationChannel(AndroidNotificationChannel channel)
         {
             self.Call("registerNotificationChannel",
@@ -182,13 +187,19 @@ namespace Unity.Notifications.Android
                 channel.CanBypassDnd,
                 channel.CanShowBadge,
                 channel.VibrationPattern,
-                (int)channel.LockScreenVisibility
+                (int)channel.LockScreenVisibility,
+                channel.Group
             );
         }
 
         public AndroidJavaObject[] GetNotificationChannels()
         {
             return self.Call<AndroidJavaObject[]>("getNotificationChannels");
+        }
+
+        public void DeleteNotificationChannelGroup(string id)
+        {
+            self.Call("deleteNotificationChannelGroup", id);
         }
 
         public void DeleteNotificationChannel(string channelId)
@@ -641,6 +652,32 @@ namespace Unity.Notifications.Android
         }
 
         /// <summary>
+        /// Register notification channel group.
+        /// </summary>
+        public static void RegisterNotificationChannelGroup(AndroidNotificationChannelGroup group)
+        {
+            if (!Initialize())
+                return;
+
+            if (string.IsNullOrEmpty(group.Id))
+                throw new Exception("Notification channel group ID is not specified.");
+            if (string.IsNullOrEmpty(group.Name))
+                throw new Exception("Notification channel group name is not specified.");
+
+            s_Jni.NotificationManager.RegisterNotificationChannelGroup(group);
+        }
+
+        /// <summary>
+        /// Delete notification channel group and all the channels in it.
+        /// </summary>
+        /// <param name="id">The ID of the group.</param>
+        public static void DeleteNotificationChannelGroup(string id)
+        {
+            if (Initialize())
+                s_Jni.NotificationManager.DeleteNotificationChannelGroup(id);
+        }
+
+        /// <summary>
         ///  Creates a notification channel that notifications can be posted to.
         ///  Notification channel settings can be changed by users on devices running Android 8.0 and above.
         ///  On older Android versions settings set on the notification channel struct will still be applied to the notification
@@ -709,6 +746,7 @@ namespace Unity.Notifications.Android
                 ch.CanShowBadge = channel.Get<bool>("canShowBadge");
                 ch.VibrationPattern = channel.Get<long[]>("vibrationPattern");
                 ch.LockScreenVisibility = channel.Get<int>("lockscreenVisibility").ToLockScreenVisibility();
+                ch.Group = channel.Get<string>("group");
 
                 channels[i] = ch;
             }